pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "nexus-migration-accelerator"
version = "1.0.0"
description = "Salesforce to Nexus CRM Migration Accelerator"
requires-python = ">=3.11"
dependencies = [
    "simple-salesforce>=1.12",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "pydantic>=2.5",
    "mcp>=1.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
]

[project.scripts]
migrate = "presentation.cli.migrate:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
